+++ src/app/components/snomed/snomed-buscar.component.html
@@ -4,7 +4,7 @@
     </div>
 </form>
 
-<!--MAESTRO DE PROBLEMAS - BD:tipoProblema-->
+<!-- MAESTRO DE PROBLEMAS - BD:tipoProblema
 <div class="menu-buscador" *ngIf="!showRefSets">
     <plex-loader class="float-right" *ngIf="loading" type="ball-pulse"></plex-loader>
     <div class="filtros d-flex">
@@ -111,14 +111,14 @@
             </li>
         </ng-container>
     </ul>
-</div>
+</div> -->
 
 <!-- refsetids -->
-<div *ngIf="showRefSets && !showFrecuentes">
+<!-- <div *ngIf="showRefSets && !showFrecuentes">
 
-    <ul class="conceptos list-unstyled">
+    <ul class="conceptos list-unstyled"> -->
         <!-- <plex-accordion> -->
-        <ng-container *ngFor="let items of arrayPorRefsets, let i = index;">
+        <!-- <ng-container *ngFor="let items of arrayPorRefsets, let i = index;">
             <li>
                 <div class="rup-card mini">
                     <div class="container rup-content">
@@ -173,6 +173,6 @@
                     </ul>
                 </div>
             </li>
-        </ng-container>
-    </ul>
-</div>
+        </ng-container> -->
+    <!-- </ul>
+</div> -->
