+++ src/app/components/snomed/snomed-buscar.component.ts
@@ -15,7 +15,7 @@ import { log } from 'util';
     templateUrl: 'snomed-buscar.component.html',
     // creamos un handler para cuando se realiza un click
     // tslint:disable-next-line:use-host-property-decorator
-    host: { '(document:click)': 'handleClick($event)' },
+    // host: { '(document:click)': 'handleClick($event)' },
     // Use to disable CSS Encapsulation for this component
     encapsulation: ViewEncapsulation.None,
     styleUrls: [
@@ -47,7 +47,7 @@ export class SnomedBuscarComponent implements OnInit, OnChanges {
     // output de informacion que devuelve el componente
     @Output() evtData: EventEmitter<any> = new EventEmitter<any>();
     @Output() tagBusqueda: EventEmitter<any> = new EventEmitter<any>();
-
+    @Output() _resultados: EventEmitter<any> = new EventEmitter<any>();
     // Output de un boolean para indicar cuando se tienen resultados de
     // busqueda o no.
     @Output() _tengoResultado: EventEmitter<any> = new EventEmitter<any>();
@@ -65,7 +65,7 @@ export class SnomedBuscarComponent implements OnInit, OnChanges {
     // ocultar lista cuando no hay resultados
     public hideLista: Boolean = false;
 
-    public resultados = [];
+    // public resultados = [];
     public elementRef;
 
     // boolean para indicar si esta cargando o no
@@ -84,17 +84,17 @@ export class SnomedBuscarComponent implements OnInit, OnChanges {
 
     public showFrecuentes = true;
 
-    /*
-    // Tipo de busqueda: hallazgos y trastornos / antecedentes / anteced. familiares
-    public tipoBusqueda: String = '';
-    */
-    public contadorSemanticTags = {
-        hallazgo: 0,
-        trastorno: 0,
-        procedimiento: 0,
-        entidadObservable: 0,
-        situacion: 0
-    };
+    // /*
+    // // Tipo de busqueda: hallazgos y trastornos / antecedentes / anteced. familiares
+    // public tipoBusqueda: String = '';
+    // */
+    // public contadorSemanticTags = {
+    //     hallazgo: 0,
+    //     trastorno: 0,
+    //     procedimiento: 0,
+    //     entidadObservable: 0,
+    //     situacion: 0
+    // };
 
     // inyectamos servicio de snomed, plex y tambien ElementRef
     // ElementRef lo utilizo para tener informacion del
@@ -208,7 +208,7 @@ export class SnomedBuscarComponent implements OnInit, OnChanges {
                 this.loading = true;
 
                 // Limpio los resultados (también se limpian los contadores)
-                this.resultados = this.resultadosAux = [];
+                // this.resultados = this.resultadosAux = [];
 
                 // buscamos
                 let apiMethod;
@@ -267,25 +267,26 @@ export class SnomedBuscarComponent implements OnInit, OnChanges {
                             let frecuentes = [];
 
                             // Esperamos que haya un resultado de más frecuentes antes de mostrar los resultados completos
-                            this.resultados = resultados;
-                            this.filtroRefSet();
-                            // En base a los resultados se arman los contadores de los filtros
-                            this.contadorSemantigTags(this.resultados);
+                            // this.resultados = resultados;
+                            this._resultados.emit(resultados);
+                            // this.filtroRefSet();
+                            // // En base a los resultados se arman los contadores de los filtros
+                            // this.contadorSemantigTags(this.resultados);
 
-                            // Hay más frecuentes?
-                            if (resultado && resultado[0] && resultado[0].frecuentes) {
+                            // Hay más frecuentes? TODO Armar los mas frecuentes en el buscador??!
+                            // if (resultado && resultado[0] && resultado[0].frecuentes) {
 
-                                // Si hay un concepto frecuente en la lista de resultados, se lo mueve al tope de la lista con Array.unshift()
-                                frecuentes = resultado[0].frecuentes.map(x => {
-                                    if (x.frecuencia != null && x.frecuencia >= 1 && this.resultados.find(c => c.conceptId === x.concepto.conceptId)) {
-                                        this.resultados.splice(this.resultados.findIndex(r => r.conceptId === x.concepto.conceptId), 1);
-                                        this.resultados.unshift(x.concepto);
-                                    }
-                                });
+                            //     // Si hay un concepto frecuente en la lista de resultados, se lo mueve al tope de la lista con Array.unshift()
+                            //     frecuentes = resultado[0].frecuentes.map(x => {
+                            //         if (x.frecuencia != null && x.frecuencia >= 1 && this.resultados.find(c => c.conceptId === x.concepto.conceptId)) {
+                            //             this.resultados.splice(this.resultados.findIndex(r => r.conceptId === x.concepto.conceptId), 1);
+                            //             this.resultados.unshift(x.concepto);
+                            //         }
+                            //     });
 
-                                // Finalmente se ordenan los más frecuentes de mayor a menor frecuencia
-                                frecuentes.sort((a, b) => b.frecuencia - a.frecuencia);
-                            }
+                            //     // Finalmente se ordenan los más frecuentes de mayor a menor frecuencia
+                            //     frecuentes.sort((a, b) => b.frecuencia - a.frecuencia);
+                            // }
 
                         });
                     }
@@ -301,116 +302,116 @@ export class SnomedBuscarComponent implements OnInit, OnChanges {
         }
     }
 
-    contadorSemantigTags(resultados): any {
-        this.contadorSemanticTags = {
-            hallazgo: 0,
-            trastorno: 0,
-            procedimiento: 0,
-            entidadObservable: 0,
-            situacion: 0
-        };
+    // contadorSemantigTags(resultados): any {
+    //     this.contadorSemanticTags = {
+    //         hallazgo: 0,
+    //         trastorno: 0,
+    //         procedimiento: 0,
+    //         entidadObservable: 0,
+    //         situacion: 0
+    //     };
 
-        let tag;
+    //     let tag;
 
-        resultados.forEach(x => {
-            tag = x.semanticTag && x.semanticTag === 'entidad observable' ? 'entidadObservable' : x.semanticTag;
-            this.contadorSemanticTags[String(tag)]++;
-        });
+    //     resultados.forEach(x => {
+    //         tag = x.semanticTag && x.semanticTag === 'entidad observable' ? 'entidadObservable' : x.semanticTag;
+    //         this.contadorSemanticTags[String(tag)]++;
+    //     });
 
-    }
+    // }
 
-    filtroBuscadorSnomed(filtro: any[], tipo = null) {
-        if (this.resultados.length >= this.resultadosAux.length && !this.loading) {
-            this.resultadosAux = this.resultados;
-        } else {
-            this.resultados = this.resultadosAux;
-        }
-        this.resultados = this.resultadosAux.filter(x => filtro.find(y => y === x.semanticTag));
-        this.tipoBusqueda = tipo ? tipo : '';
-        this.filtroActual = tipo ? ['planes'] : filtro;
-        this.esFiltroActual = this.getFiltroActual(filtro);
-        return this.resultados;
-    }
+    // filtroBuscadorSnomed(filtro: any[], tipo = null) {
+    //     if (this.resultados.length >= this.resultadosAux.length && !this.loading) {
+    //         this.resultadosAux = this.resultados;
+    //     } else {
+    //         this.resultados = this.resultadosAux;
+    //     }
+    //     this.resultados = this.resultadosAux.filter(x => filtro.find(y => y === x.semanticTag));
+    //     this.tipoBusqueda = tipo ? tipo : '';
+    //     this.filtroActual = tipo ? ['planes'] : filtro;
+    //     this.esFiltroActual = this.getFiltroActual(filtro);
+    //     return this.resultados;
+    // }
 
-    // :joy:
-    getFiltroActual(filtro: any[]) {
-        return this.filtroActual.join('') === filtro.join('');
-    }
+    // // :joy:
+    // getFiltroActual(filtro: any[]) {
+    //     return this.filtroActual.join('') === filtro.join('');
+    // }
 
-    /**
-     * Handler para cuando se ejecuta un click en el documento.
-     * @param event  Click event
-     * @returns      Void
-     */
-    handleClick(event): void {
-        // buscamos que elemento fue clickeado
-        let clickedComponent = event.target;
+    // /**
+    //  * Handler para cuando se ejecuta un click en el documento.
+    //  * @param event  Click event
+    //  * @returns      Void
+    //  */
+    // handleClick(event): void {
+    //     // buscamos que elemento fue clickeado
+    //     let clickedComponent = event.target;
 
-        // creamos una bandera para saber si pertenece a este componente
-        let inside = false;
+    //     // creamos una bandera para saber si pertenece a este componente
+    //     let inside = false;
 
-        // loopeamos
-        do {
-            // si hice click dentro del codigo html del componente
-            // entonces indico que estoy adentro (inside = true)
-            // y no oculto la lista de resultados
-            if (clickedComponent === this.elementRef.nativeElement) {
-                inside = true;
+    //     // loopeamos
+    //     do {
+    //         // si hice click dentro del codigo html del componente
+    //         // entonces indico que estoy adentro (inside = true)
+    //         // y no oculto la lista de resultados
+    //         if (clickedComponent === this.elementRef.nativeElement) {
+    //             inside = true;
 
-                this.hideLista = false;
-            }
+    //             this.hideLista = false;
+    //         }
 
-            // info de que componente hice clic
-            clickedComponent = clickedComponent.parentNode;
-        } while (clickedComponent);
+    //         // info de que componente hice clic
+    //         clickedComponent = clickedComponent.parentNode;
+    //     } while (clickedComponent);
 
-        // si no estamos en el componente, limpiamos lista de problemas
-        if (!inside && !this._draggable) {
-            // this.resultados = [];
-            // this.hideLista = true;
-            // this.searchTerm = '';
-        }
-    }
+    //     // si no estamos en el componente, limpiamos lista de problemas
+    //     if (!inside && !this._draggable) {
+    //         // this.resultados = [];
+    //         // this.hideLista = true;
+    //         // this.searchTerm = '';
+    //     }
+    // }
 
-    // si hago clic en un concepto, entonces lo devuelvo
-    seleccionarConcepto(concepto) {
-        // this.resultados = this.resultadosAux = [];
-        // this.searchTerm = '';
-        // this.contadorSemanticTags = {
-        //     hallazgo: 0,
-        //     trastorno: 0,
-        //     procedimiento: 0,
-        //     entidadObservable: 0,
-        //     situacion: 0
-        // };
-        this.tagBusqueda.emit(this.filtroActual);
-        this.evtData.emit(concepto);
-    }
+    // // si hago clic en un concepto, entonces lo devuelvo
+    // seleccionarConcepto(concepto) {
+    //     // this.resultados = this.resultadosAux = [];
+    //     // this.searchTerm = '';
+    //     // this.contadorSemanticTags = {
+    //     //     hallazgo: 0,
+    //     //     trastorno: 0,
+    //     //     procedimiento: 0,
+    //     //     entidadObservable: 0,
+    //     //     situacion: 0
+    //     // };
+    //     this.tagBusqueda.emit(this.filtroActual);
+    //     this.evtData.emit(concepto);
+    // }
 
-    filtroRefSet() {
-        let conceptos = {
-            Hallazgos: ['hallazgo', 'situacion'],
-            Trastornos: ['trastorno'],
-            Procedimientos: ['procedimiento', 'entidad observable'],
-            Planes: ['procedimiento']
-        };
-        this.arrayPorRefsets = [];
-        Object.keys(this.servicioPrestacion.refsetsIds).forEach(k => {
-            let nombre = k.replace(/_/g, ' ');
-            this.arrayPorRefsets.push({ nombre: nombre, valor: this.resultados.filter(x => x.refsetIds.find(item => item === this.servicioPrestacion.refsetsIds[k])) });
-        });
-        Object.keys(conceptos).forEach(c => {
-            this.arrayPorRefsets.push({ nombre: c, valor: this.filtroBuscadorSnomed(conceptos[c]) });
-        });
-    }
+    // filtroRefSet() {
+    //     let conceptos = {
+    //         Hallazgos: ['hallazgo', 'situacion'],
+    //         Trastornos: ['trastorno'],
+    //         Procedimientos: ['procedimiento', 'entidad observable'],
+    //         Planes: ['procedimiento']
+    //     };
+    //     this.arrayPorRefsets = [];
+    //     Object.keys(this.servicioPrestacion.refsetsIds).forEach(k => {
+    //         let nombre = k.replace(/_/g, ' ');
+    //         this.arrayPorRefsets.push({ nombre: nombre, valor: this.resultados.filter(x => x.refsetIds.find(item => item === this.servicioPrestacion.refsetsIds[k])) });
+    //     });
+    //     Object.keys(conceptos).forEach(c => {
+    //         this.arrayPorRefsets.push({ nombre: c, valor: this.filtroBuscadorSnomed(conceptos[c]) });
+    //     });
+    // }
 
-    desplegar(i, nombre) {
-        if (this.showContent === nombre) {
-            this.showContent = null;
-        } else {
-            this.showContent = nombre;
-        }
+    // desplegar(i, nombre) {
+    //     if (this.showContent === nombre) {
+    //         this.showContent = null;
+    //     } else {
+    //         this.showContent = nombre;
+    //     }
 
-    }
+    // }
 
 }
